Use a more dynamic approach for the navigation generation.
This allows us to use the same menu items in both Mobile and Desktop.

We could probably even migrate this to a JSON format, if that helps…eg. use `title_from_page: 'pricing'` to execute on the fly, instead of `locale_page(page: 'pricing', locale_obj: locale_obj)[:page_title]` in instantiation..