When fetching landing pages, use the locale so we get the right title.
Previously, we would do a `get_landing_page('faq')` and we'd get back the global page, so the title would be "Frequently Asked Questions | Sharesight".  The same would be for any locale.  We want "Frequently Asked Questions | Sharesight NZ" (etc).

This is visible in the Footer currently and was problematic in the Header as well.

 - Minor cleanup to the `landing_pages_collection` as we can `.reject` first if we de-tuple it with `.values`.  Should be more performant and readable.
 - A fair bit of test coverage!