Remove Bugsnag
We haven't had any Bugsnags in the past 60 days so it's either that this isn't working correctly or it's not valid.  Given we don't actively maintain this and we're cleaning up Bugsnag, we should clean this up.

The Github Action which sent releases appeared to no longer be working as the last one was >2 years old.  The
`secrets.BUGSNAG_API_KEY` should be deleted.

We may however want Bugsnag in the new Gatsby website as it uses JavaScript a lot more!